requirements.txt
llama_index
google-generativeai
llama-index-llms-gemini
pypdf
python-dotenv
IPython
llama-index-embeddings-gemini
streamlit